finish project object to lines

+++ include/closestPairInLine.hpp
@@ -5,7 +5,6 @@
 #define MEDIAN_FLAG 0
 #define PIVOT_FLAG 1
 
-
-std::pair<int, int> getClosestPairInLine(std::vector<int> v, int flag);
+std::pair<int, int> getClosestPairInLine(std::vector<double> v, int flag);
 
 #endif
